refactor(avatar): extract identity helper and add daemon notification for CLI (#27804)
* refactor(avatar): extract identity helper and add daemon notification for CLI

* fix(ci): regenerate openapi spec and add knip entry for daemon-avatar-client
